# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Driver shift diary",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/days": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List Days",
        "operationId": "list_days_api_days_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayInfo"
                  },
                  "title": "Response List Days Api Days Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List Trips",
        "operationId": "list_trips_api_trips_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  },
                  "title": "Response List Trips Api Trips Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Add Trip",
        "operationId": "add_trip_api_trips_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/summary": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Day Summary",
        "operationId": "day_summary_api_summary_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}": {
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Update Trip",
        "operationId": "update_trip_api_trips__trip_id__patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Delete Trip",
        "operationId": "delete_trip_api_trips__trip_id__delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/drivers/{driver_id}/trips": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Driver Trips",
        "operationId": "driver_trips_api_admin_drivers__driver_id__trips_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  },
                  "title": "Response Driver Trips Api Admin Drivers  Driver Id  Trips Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/drivers/{driver_id}/trips/{trip_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Driver Trip",
        "operationId": "update_driver_trip_api_admin_drivers__driver_id__trips__trip_id__patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Driver Trip",
        "operationId": "delete_driver_trip_api_admin_drivers__driver_id__trips__trip_id__delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DayInfo": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "net": {
            "type": "integer",
            "title": "Net"
          }
        },
        "type": "object",
        "required": [
          "date",
          "count",
          "net"
        ],
        "title": "DayInfo",
        "description": "Short per-day entry for the day navigation panel: a day with at least one shift."
      },
      "DaySummary": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "revenue": {
            "type": "integer",
            "title": "Revenue"
          },
          "commission": {
            "type": "integer",
            "title": "Commission"
          },
          "net": {
            "type": "integer",
            "title": "Net"
          },
          "cash": {
            "$ref": "#/components/schemas/PaymentBreakdown"
          },
          "card": {
            "$ref": "#/components/schemas/PaymentBreakdown"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "shifts": {
            "type": "integer",
            "title": "Shifts"
          }
        },
        "type": "object",
        "required": [
          "count",
          "revenue",
          "commission",
          "net",
          "cash",
          "card",
          "date",
          "shifts"
        ],
        "title": "DaySummary",
        "description": "Totals of all trips in the shifts that started on this local day."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentBreakdown": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "amount": {
            "type": "integer",
            "title": "Amount",
            "default": 0
          }
        },
        "type": "object",
        "title": "PaymentBreakdown"
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "cash",
          "card"
        ],
        "title": "PaymentMethod"
      },
...
      "Trip": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "shift_id": {
            "type": "integer",
            "title": "Shift Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "commission": {
            "type": "integer",
            "title": "Commission"
          },
          "commission_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shift_id",
          "start",
          "end",
          "amount",
          "payment",
          "commission"
        ],
        "title": "Trip",
        "description": "A stored trip."
      },
      "TripIn": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "shift_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Shift Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "amount": {
            "type": "integer",
            "maximum": 500000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Trip amount, KZT"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "commission": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission",
            "description": "Commission, KZT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "shift_id",
          "start",
          "end",
          "amount",
          "payment"
        ],
        "title": "TripIn",
        "description": "Trip as sent by a client. The id is optional."
      },
      "TripPatch": {
        "properties": {
          "shift_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Id"
          },
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "amount": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "payment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "commission": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TripPatch",
        "description": "Changes to a stored trip; omitted fields keep their values. The id never changes."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}